Refactor shape class to encapsulate the shape data

+++ ModelGraph/ModelGraph/Controls/Shapes/Polygon.cs
@@ -24,6 +24,6 @@ namespace ModelGraph.Controls
             }
         }
 
-        protected override (byte min, byte max) MinMaxDimension => (2, 8);
+        protected override (int min, int max) MinMaxDimension => (2, 8);
     }
 }
